+++ test/support/test_repo/migrations/20221011201645_install_carbonite_alternate_test_schema.exs
@@ -4,13 +4,7 @@ defmodule Carbonite.TestRepo.Migrations.InstallCarboniteAlternateTestSchema do
   use Ecto.Migration
 
   def up do
-    Carbonite.Migrations.up(1, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(2, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(3, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(4, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(5, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(6, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.up(7, carbonite_prefix: "alternate_test_schema")
+    Carbonite.Migrations.up(1..7, carbonite_prefix: "alternate_test_schema")
 
     Carbonite.Migrations.create_trigger(:rabbits, carbonite_prefix: "alternate_test_schema")
 
@@ -26,12 +20,6 @@ defmodule Carbonite.TestRepo.Migrations.InstallCarboniteAlternateTestSchema do
   def down do
     Carbonite.Migrations.drop_trigger(:rabbits, carbonite_prefix: "alternate_test_schema")
 
-    Carbonite.Migrations.down(7, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(6, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(5, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(4, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(3, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(2, carbonite_prefix: "alternate_test_schema")
-    Carbonite.Migrations.down(1, carbonite_prefix: "alternate_test_schema")
+    Carbonite.Migrations.down(7..1, carbonite_prefix: "alternate_test_schema")
   end
 end